Don't wrap AtlasIndexProvider with ResourceCache
